+++ android/src/main/java/com/adjust/sdk/flutter/AdjustSdk.java
@@ -423,6 +423,7 @@ public class AdjustSdk implements FlutterPlugin, MethodCallHandler {
                                 adjustAttribution.costAmount.toString() : "");
                         adjustAttributionMap.put("costCurrency", adjustAttribution.costCurrency);
                         adjustAttributionMap.put("fbInstallReferrer", adjustAttribution.fbInstallReferrer);
+                        adjustAttributionMap.put("jsonResponse", adjustAttribution.jsonResponse);
                         if (channel != null) {
                             channel.invokeMethod(dartMethodName, adjustAttributionMap);
                         }
@@ -734,6 +735,7 @@ public class AdjustSdk implements FlutterPlugin, MethodCallHandler {
                         attribution.costAmount.toString() : "");
                 adjustAttributionMap.put("costCurrency", attribution.costCurrency);
                 adjustAttributionMap.put("fbInstallReferrer", attribution.fbInstallReferrer);
+                adjustAttributionMap.put("jsonResponse", attribution.jsonResponse);
                 result.success(adjustAttributionMap);
             }
         });