+++ ios/Classes/AdjustSdkDelegate.m
@@ -118,8 +118,15 @@ static NSString *dartSkanUpdatedCallback = nil;
         @"clickLabel",
         @"costType",
         @"costAmount",
-        @"costCurrency"
+        @"costCurrency",
+        @"jsonResponse"
     };
+    NSData *dataJsonResponse = [NSJSONSerialization dataWithJSONObject:attribution.jsonResponse
+                                                               options:0
+                                                                 error:nil];
+    NSString *stringJsonResponse = [[NSString alloc] initWithBytes:[dataJsonResponse bytes]
+                                                            length:[dataJsonResponse length]
+                                                          encoding:NSUTF8StringEncoding];
     id values[] = {
         [self getValueOrEmpty:[attribution trackerToken]],
         [self getValueOrEmpty:[attribution trackerName]],
@@ -130,12 +137,13 @@ static NSString *dartSkanUpdatedCallback = nil;
         [self getValueOrEmpty:[attribution clickLabel]],
         [self getValueOrEmpty:[attribution costType]],
         [self getNumberValueOrEmpty:[attribution costAmount]],
-        [self getValueOrEmpty:[attribution costCurrency]]
+        [self getValueOrEmpty:[attribution costCurrency]],
+        [self getValueOrEmpty:stringJsonResponse]
     };
     NSUInteger count = sizeof(values) / sizeof(id);
     NSDictionary *attributionMap = [NSDictionary dictionaryWithObjects:values
-                                                                     forKeys:keys
-                                                                       count:count];
+                                                               forKeys:keys
+                                                                 count:count];
     [self.channel invokeMethod:dartAttributionCallback arguments:attributionMap];
 }