+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+### Version 5.1.0 (25th February 2025)
+#### Added
+- Added `jsonResponse` field (JSON string) to `AdjustAttribution` where every key-value pair sent by the backend as part of the attribution response can be found.
+
+#### Native SDKs
+- [iOS@v5.1.1][ios_sdk_v5.1.1]
+- [Android@v5.1.0][android_sdk_v5.1.0]
+
+---
+
 ### Version 5.0.4 (6th February 2025)
 #### Added
 - Added sending of the additional parameter to improve troubleshooting of `SKAdNetwork` related issues.
@@ -515,7 +525,7 @@ In case you were using beta version of the SDK v5, please switch to the official
 [ios_sdk_v4.38.4]: https://github.com/adjust/ios_sdk/tree/v4.38.4
 [ios_sdk_v5.0.0]: https://github.com/adjust/ios_sdk/tree/v5.0.0
 [ios_sdk_v5.0.1]: https://github.com/adjust/ios_sdk/tree/v5.0.1
-[ios_sdk_v5.1.0]: https://github.com/adjust/ios_sdk/tree/v5.1.0
+[ios_sdk_v5.1.1]: https://github.com/adjust/ios_sdk/tree/v5.1.1
 
 [android_sdk_v4.17.0]: https://github.com/adjust/android_sdk/tree/v4.17.0
 [android_sdk_v4.18.0]: https://github.com/adjust/android_sdk/tree/v4.18.0